GWin32AppInfo: Store UWP AppUserModelId for a handler
Could be used later on to launch UWP apps.

UWP apps need verbs, but we don't have any code to get UWP app
info yet.

+++ gobject/tests/properties.c
@@ -554,7 +554,7 @@ properties_get_property (void)
     { "bar", G_TYPE_INVALID, G_VALUE_INIT },
     { "bar", G_TYPE_STRING, G_VALUE_INIT },
   };
-  gsize i;
+  int i;
 
   g_test_summary ("g_object_get_property() accepts uninitialized, "
                   "initialized, and transformable values");
